ALIEN: Stand the character where the room says, in the rooms that say it inline
Fifty-three placements over twenty-two rooms were missing from the
opening table, so loadRoom fell back to the room's first walk node. The
crossroads put him at 9,116 -- the far left, on the mansion side --
however he came in: the path back was the one he stood on, and anything
the node ring could not see from that corner made the router lap and
settle for a route across blocked ground, in front of the scenery.

The coordinates are longs whose high word is not always zero, so the low
word is read as signed: the shore walks him in from above the plate and
the diving area from left of it.

+++ engines/alien/script.cpp
@@ -502,11 +502,13 @@ void RoomScript::runEffect(const ScriptEffect &original) {
 		break;
 
 	case kOpCharPlace:
-		// The coordinates are longs in the original and their high words are
-		// always zero, which is why only the low ones are read here.
+		// The coordinates are longs in the original, and the high word is not
+		// always zero: the shore and the diving area walk the character in from
+		// off the plate, above it or left of it, so the low word is read as
+		// signed and the high word is only there to say which.
 		_placed = true;
-		_placeX = effect.args[1];
-		_placeY = effect.args[3];
+		_placeX = (int16)effect.args[1];
+		_placeY = (int16)effect.args[3];
 		_placeFacing = effect.args[4];
 		break;
